+++ CHANGELOG.md
@@ -9,6 +9,7 @@
 * Coordinates: Added `AreReal` property to confirm `X` and `Y` are finite
 * Crosshair: Added `X` and `Y` properties to compliment `Position`
 * Axes: Removed `Plot.Axes` list to encourage interaction with `Plot.YAxes` and `Plot.XAxes`
+* Layout: Created `Plot.Layout.Frameless()` to hide axes and allow the data area to fill the figure
 
 ## ScottPlot 5.0.11-beta
 _Published on [NuGet](https://www.nuget.org/profiles/ScottPlot) on 2023-12-29_